pyproject.toml
[project]
name = "cat-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.8",
    "duckdb>=1.2.2",
    "av>=14.4.0",
    "pillow>=11.1.0",
    "ultralytics>=8.3.0",
    "numpy>=2.2.0",
    "imagehash>=4.3.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
